Move `_vec` to src
