+++ experimental/GModule/GModule.jl
@@ -8,7 +8,7 @@ import Hecke: data
 #   first does a "restriction of scalars" or blow up with the rep mat
 #   second tries to conjugate down to k
 
-import Oscar: gmodule, GAPWrap
+import Oscar: _vec, gmodule, GAPWrap
 import Oscar.GrpCoh: MultGrp, MultGrpElem
 
 import AbstractAlgebra: Group, Module
@@ -1442,22 +1442,6 @@ end
 
 #TODO: cover all finite fields
 #      make the Modules work
-
-#to bypass the vec(collect(M)) which copies twice
-function _vec(M::Generic.Mat)
-  return vec(M.entries)
-end
-
-function _vec(M::MatElem)
-  r = elem_type(base_ring(M))[]
-  sizehint!(r, nrows(M) * ncols(M))
-  for j=1:ncols(M)
-    for i=1:nrows(M)
-      push!(r, M[i, j])
-    end
-  end
-  return r
-end
 
 function Oscar.simplify(C::GModule{<:Any, <:AbstractAlgebra.FPModule{QQFieldElem}})
   return gmodule(QQ, Oscar.simplify(gmodule(ZZ, C))[1])
