+++ src/Modules/missing_functionality.jl
@@ -84,3 +84,18 @@ function (==)(M1::SubquoModule{T}, M2::SubquoModule{T}) where {T<:AbsLocalizedRi
 end
 
 
+#to bypass the vec(collect(M)) which copies twice
+function _vec(M::Generic.Mat)
+  return vec(M.entries)
+end
+
+function _vec(M::MatElem)
+  r = elem_type(base_ring(M))[]
+  sizehint!(r, nrows(M) * ncols(M))
+  for j=1:ncols(M)
+    for i=1:nrows(M)
+      push!(r, M[i, j])
+    end
+  end
+  return r
+end
